requirements.txt
aiofiles==0.8.0
allure-pytest==2.9.45
allure-python-commons==2.9.45
altgraph==0.17.3
app==0.0.1
asgiref==3.5.1
async-generator==1.10
async-timeout==4.0.2
atomicwrites==1.4.0
attrs==21.2.0
auto-py-to-exe==2.23.1
blinker==1.4
bottle==0.12.2
Brotli==1.0.9
certifi==2021.10.8
cffi==1.15.0
chardet==4.0.0
charset-normalizer==2.0.7
click==8.1.3
colorama==0.4.4
colorlog==6.6.0
crypto==1.4.1
cryptography==36.0.0
Deprecated==1.2.13
DingtalkChatbot==1.5.3
Eel==0.14.0
et-xmlfile==1.1.0
execnet==1.9.0
Faker==9.8.3
Flask==2.0.3
future==0.18.2
gevent==22.10.2
gevent-websocket==0.10.1
greenlet==2.0.0.post0
h11==0.13.0
h2==4.1.0
hpack==4.0.0
httptools==0.4.0
hyperframe==6.0.1
idna==3.3
iniconfig==1.1.0
itchat==1.3.10
itsdangerous==2.1.2
Jinja2==3.1.2
jsonpath==0.82
kaitaistruct==0.9
ldap3==2.9.1
lib==4.0.0
lxml==4.8.0
MarkupSafe==2.1.1
mitmproxy==8.1.1
msgpack==1.0.3
multidict==6.0.2
Naked==0.1.32
numpy==1.22.4
openpyxl==3.0.9
option==2.1.0
outcome==1.1.0
packaging==21.3
pandas==1.4.2
passlib==1.7.4
pefile==2022.5.30
pluggy==1.0.0
protobuf==3.19.4
publicsuffix2==2.20191221
py==1.11.0
pyasn1==0.4.8
pycparser==2.21
pydantic==1.8.2
pyDes==2.0.1
pydivert==2.1.0
pyinstaller==5.6.2
pyinstaller-hooks-contrib==2022.12
PyMySQL==1.0.2
pyOpenSSL==21.0.0
pyparsing==3.0.6
pyperclip==1.8.2
pypinyin==0.47.1
pypng==0.0.21
PyQRCode==1.2.1
PySocks==1.7.1
pytest==7.1.2
pytest-base-url==2.0.0
pytest-forked==1.3.0
pytest-html==3.1.1
pytest-metadata==2.0.1
pytest-ordering==0.6
pytest-rerunfailures==10.2
pytest-xdist==2.4.0
python-dateutil==2.8.2
pytz==2022.1
pywin32==304
pywin32-ctypes==0.2.0
PyYAML==5.4.1
redis==4.3.4
requests==2.26.0
requests-toolbelt==0.9.1
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.6
sanic==22.3.1
sanic-routing==22.3.0
selenium==4.1.5
shellescape==3.8.1
six==1.16.0
sniffio==1.2.0
sortedcontainers==2.4.0
text-unidecode==1.3
toml==0.10.2
tomli==2.0.1
tornado==6.1
trio==0.20.0
trio-websocket==0.9.2
typing_extensions==4.4.0
urllib3==1.26.7
urwid==2.1.2
webdriver-manager==3.5.4
websockets==10.3
Werkzeug==2.1.2
whichcraft==0.6.1
wrapt==1.14.1
wsproto==1.1.0
xlrd==2.0.1
xlutils==2.0.0
xlwings==0.27.7
xlwt==1.3.0
zope.event==4.5.0
zope.interface==5.5.1
zstandard==0.17.0